world_model/multi_modal/requirements.txt
# Multi-modal world model dependencies (image editing)
torch>=2.0.0
diffusers
Pillow
transformers
accelerate
